refactor(company-setting): move form components to new folder structure
