--- src/components/sections/company-setting/index.ts
@@ -1,4 +1,0 @@
-export * from "./company-logo-section";
-export * from "./section-title";
-export * from "./setting-tab-nav";
-export * from "./tag-input";
