+++ src/components/sections/settings/index.ts
@@ -1,0 +1,3 @@
+export * from "./logo-section";
+export * from "./section-title";
+export * from "./recruiter";
